# swagger.yaml
swagger: '2.0'
info:
  title: "Спортик и точка"
  description: "Описание API"
  version: "1.0"
components:
  schemas:
    NotFoundException:
      type: object
      properties:
        success:
          type: boolean
          description: "Результат операции"
        error:
          type: string
          description: "Сообщение о том, кто или что не найдено"
    Unauthorized:
      type: object
      properties:
        success:
          type: boolean
          description: "Результат операции"
        error:
          type: string
          description: "Сообщение об ошибке"
    BadRequest:
      type: object
      properties:
        success:
          type: boolean
          description: "Результат операции"
        error:
          type: string
          description: "Сообщение об ошибке"
    Conflict:
      type: object
      properties:
        free:
          type: boolean
          description: "Проверка доступности email"
    InvalidRoleException:
      type: object
      properties:
        success:
          type: boolean
          description: "Результат операции"
        error:
          type: string
          description: "Описание ошибки"
...
    graph_data:
      type: object
      required:
        - date
        - users_with_premium
        - users_without_premium
      properties:
        date:
          type: string
          description: Дата активности в формате "гггг-мм-дд"
        users_with_premium:
          type: integer
          description: Количество премиум-пользователей
        users_without_premium:
          type: integer
          description: Количество обычных пользователей
...
    PremiumActivityResponse:
      type: object
      required:
        - id
        - activity_type
        - image
        - date
        - avg_speed
        - distance_in_meters
        - duration
        - calories_burned
      properties:
        id:
          type: integer
          description: "ID активности"
        activity_type:
          type: string
          description: "Тип активности (running, cycling, swimming)"
        image:
          type: string
          description: "Отображение пути на карте"
        date:
          type: string
          format: date
          description: "Дата активности в формате 'гггг-мм-дд'"
        avg_speed:
          type: number
          description: "Средняя скорость"
        distance_in_meters:
          type: integer
          description: "Расстояние в метрах"
        duration:
          type: integer
          description: "Продолжительность в секундах"
        calories_burned:
          type: integer
          description: "Сожженные калории"
    AchievementResponse:
      type: object
      required:
        - id
        - name
        - image
        - distance
      properties:
        id:
          type: integer
          description: ID достижения
        name:
          type: string
          description: Название достижения
        image:
          type: string
          description: Картинка достижения
        distance:
          type: integer
          description: Расстояние для получения ачивки
...
paths:
  /user/register:
    post:
      tags:
        - "Пользователи"
      summary: "Создать новый аккаунт пользователя"
      consumes:
        - multipart/form-data
      parameters:
        - name: "email"
          in: "header"
          required: true
          type: "string"
          description: "Email пользователя"
        - name: avatar
          in: formData
          description: The uploaded file data
          required: false
          type: file
        - name: name
          in: formData
          type: string
          required: true
          description: Имя пользователя
        - name: birthday
          in: formData
          type: string
          required: true
          description: Дата рождения
        - name: weight
          in: formData
          type: number
          format: float
          required: true
          description: Вес
        - name: phone
          in: formData
          type: string
          required: true
          description: Телефон
        - name: password_hash
          in: formData
          type: string
          required: true
          description: Пароль
      responses:
        '201':
          description: "Успешная регистрация"
          schema:
            id: "RegisterResponse"
            properties:
              success:
                type: "boolean"
                description: "Результат регистрации"
              access_token:
                type: "string"
                description: "Токен доступа пользователя"
              user_id:
                type: "integer"
                description: "ID пользователя"
        '400':
          description: "Ошибка введенных данных"
          schema:
            type: object
            $ref: '#/components/schemas/BadRequest'
        '409':
          description: "Ошибка, если email уже используется"
          schema:
            type: object
            $ref: '#/components/schemas/Conflict'
  /admin/register:
    post:
      tags:
        - "Администраторы"
      summary: "Создать новый аккаунт администратора"
      consumes:
        - multipart/form-data
      parameters:
        - name: "email"
          in: "header"
          required: true
          type: "string"
          description: "Email администратора"
        - name: avatar
          in: formData
          description: The uploaded file data
          required: false
          type: file
        - name: name
          in: formData
          type: string
          required: true
          description: Имя администратора
        - name: birthday
          in: formData
          type: string
          required: true
          description: Дата рождения
        - name: phone
          in: formData
          type: string
          required: true
          description: Телефон
        - name: password_hash
          in: formData
          type: string
          required: true
          description: Пароль
      responses:
        '201':
          description: "Успешная регистрация"
          schema:
            id: "RegisterAdminResponse"
            properties:
              success:
                type: "boolean"
                description: "Результат регистрации"
              access_token:
                type: "string"
                description: "Токен доступа администратора"
              admin_id:
                type: "integer"
                description: "ID администратора"
        '400':
          description: "Ошибка введенных данных"
          schema:
            type: object
            $ref: '#/components/schemas/BadRequest'
        '409':
          description: "Ошибка, если email уже используется"
          schema:
            type: object
            $ref: '#/components/schemas/Conflict'
...
  /admin/action:
    put:
      tags:
        - "Администраторы"
      summary: "Заблокировать, разблокировать, забрать премиум"
      parameters:
        - name: Authorization
          in: header
          required: true
          type: string
          description: Токен доступа администратора
        - name: request_data
          in: body
          required: true
          schema:
            id: AdminModifyActionRequest
            properties:
              user_id:
                type: integer
                description: ID пользователя
              action:
                type: string
                enum: [ block, unblock, revoke_premium ]
                description: Выбранное действие
      responses:
        200:
          description: Успешная модификация статуса пользователя
          schema:
            id: AdminModifyActionResponse
            properties:
              success:
                type: boolean
                description: Результат действия
              action:
                type: string
                description: Выполненное действие
        400:
          description: "Ошибка введенных данных"
          schema:
            type: object
            $ref: '#/components/schemas/BadRequest'
        401:
          description: Токен некорректный
          schema:
            type: object
            $ref: '#/components/schemas/Unauthorized'
        403:
          description: Доступ запрещен (Forbidden)
          schema:
            type: object
            $ref: '#/components/schemas/InvalidRoleException'
        404:
          description: Пользователь или администратор не найдены
          schema:
            type: object
            $ref: '#/components/schemas/NotFoundException'
        409:
          description: Действие недоступно
          schema:
            type: object
            $ref: '#/components/schemas/Conflict'
    post:
      tags:
        - "Администраторы"
      summary: "Выдать премиум-подписку"
      parameters:
        - name: Authorization
          in: header
          required: true
          type: string
          description: Токен доступа администратора
        - name: request_data
          in: body
          required: true
          schema:
            id: AdminGrantPremiumRequest
            properties:
              user_id:
                type: integer
                description: ID пользователя
      responses:
        200:
          description: Успешная выдача премиума пользователю
          schema:
            id: AdminGrantPremiumResponse
            properties:
              success:
                type: boolean
                description: Результат действия
              action:
                type: string
                description: Выполненное действие
        401:
          description: Токен некорректный
          schema:
            type: object
            $ref: '#/components/schemas/Unauthorized'
        403:
          description: Доступ запрещен (Forbidden)
          schema:
            type: object
            $ref: '#/components/schemas/InvalidRoleException'
        404:
          description: Пользователь или администратор не найдены
          schema:
            type: object
            $ref: '#/components/schemas/NotFoundException'
        409:
          description: Действие недоступно
          schema:
            type: object
            $ref: '#/components/schemas/Conflict'
...
  /user/profile:
    get:
      tags:
        - "Пользователи"
      summary: "Получить профиль пользователя"
      description: "Возвращает данные для профиля пользователя"
      parameters:
        - name: Authorization
          in: header
          required: true
          type: string
          description: Токен доступа пользователя
        - name: period
          in: query
          required: true
          type: string
          enum: [ week, month, year, all_time ]
          description: Период времени для статистики
      responses:
        200:
          description: Успешный запрос
          schema:
            id: UserProfileResponse
            properties:
              name:
                type: string
                description: Имя пользователя
              image:
                type: string
                description: Фото пользователя
              rating:
                type: integer
                description: Место пользователя в рейтинге
              statistics:
                type: object
                description: Статистика пользователя
                properties:
                  total_distance_in_meters:
                    type: integer
                    description: Общее расстояние в метрах
                  total_time:
                    type: integer
                    description: Общее время
                  total_calories:
                    type: integer
                    description: Общее количество калорий
              achievements:
                type: array
                items:
                  $ref: '#/components/schemas/AchievementResponse'
        401:
          description: Токен некорректный
          schema:
            type: object
            $ref: '#/components/schemas/Unauthorized'
        403:
          description: Доступ запрещен (Forbidden)
          schema:
            type: object
            $ref: '#/components/schemas/InvalidRoleException'
        404:
          description: Ошибка, если пользователь или статистика не найдены
          schema:
            type: object
            $ref: '#/components/schemas/NotFoundException'
  /admin/profile:
    get:
      tags:
        - "Администраторы"
      summary: "Получить профиль администратора"
      description: "Возвращает данные для профиля администратора"
      parameters:
        - name: Authorization
          in: header
          required: true
          type: string
          description: Токен доступа пользователя
      responses:
        200:
          description: Успешный запрос
          schema:
            id: AdminProfileResponse
            properties:
              name:
                type: string
                description: Имя админа
              image:
                type: string
                description: Фото админа
        401:
          description: Токен некорректный
          schema:
            type: object
            $ref: '#/components/schemas/Unauthorized'
        403:
          description: Доступ запрещен (Forbidden)
          schema:
            type: object
            $ref: '#/components/schemas/InvalidRoleException'
        404:
          description: Ошибка, если администратор не найден
          schema:
            type: object
            $ref: '#/components/schemas/NotFoundException'
...
  /user/premium/statistics:
    get:
      tags:
        - "Статистика"
      summary: "Получить статистику для премиум-пользователя"
      description: "Возвращает данные для премиум-статистики"
      parameters:
        - name: Authorization
          in: header
          required: true
          type: string
          description: "Токен доступа пользователя"
        - name: period
          in: query
          required: true
          type: string
          enum: [ week, month, year, all_time ]
          description: "Период времени для статистики"
      responses:
        200:
          description: "Успешный запрос"
          schema:
            id: PremiumStatisticsResponse
            properties:
              total_distance_in_meters:
                type: integer
                description: "Общее расстояние в метрах"
              total_time:
                type: integer
                description: "Общее время"
              total_calories:
                type: integer
                description: "Общее количество калорий"
              avg_speed:
                type: number
                format: float
                description: "Средняя скорость"
              activities:
                type: array
                items:
                  $ref: '#/components/schemas/PremiumActivityResponse'
        401:
          description: "Токен некорректный"
          schema:
            type: object
            $ref: '#/components/schemas/Unauthorized'
        403:
          description: Доступ запрещен (Forbidden)
          schema:
            type: object
            $ref: '#/components/schemas/InvalidRoleException'
        404:
          description: "Ошибка, если пользователь или статистика не найдены"
          schema:
            type: object
            $ref: '#/components/schemas/NotFoundException'
  /admin/statistics:
    get:
      tags:
        - "Статистика"
      summary: "Получить статистику для администраторов"
      description: "Возвращает данные для статистики администратора"
      parameters:
        - name: Authorization
          in: header
          required: true
          type: string
          description: Токен доступа администратора
        - name: period
          in: query
          required: true
          type: string
          enum: [ week, month, year, all_time ]
          description: Период времени для статистики
      responses:
        200:
          description: Успешный запрос
          schema:
            id: AdminStatisticsRequest
            properties:
              total_users:
                type: integer
                description: Общее количество пользователей приложения
              premium_users:
                type: integer
                description: Общее количество премиум-пользователей приложения
              graph_data:
                type: array
                items:
                  $ref: '#/components/schemas/graph_data'
        401:
          description: Токен некорректный
          schema:
            type: object
            $ref: '#/components/schemas/Unauthorized'
        403:
          description: Доступ запрещен (Forbidden)
          schema:
            type: object
            $ref: '#/components/schemas/InvalidRoleException'
        404:
          description: Ошибка, если пользователь или статистика не найдены
          schema:
            type: object
            $ref: '#/components/schemas/NotFoundException'
...